+++ BookStore/BookStore.Business.Services/BookStore.Business.Services.csproj
@@ -67,9 +67,9 @@
       <Project>{c08dce29-b09c-4377-bc05-d14bc403c71d}</Project>
       <Name>BookStore.Business</Name>
     </ProjectReference>
-    <ProjectReference Include="..\BookStore.Data.EntityFramework\BookStore.Data.EntityFramework.csproj">
-      <Project>{77bebff9-6b25-4999-a333-8cb5e74b2b82}</Project>
-      <Name>BookStore.Data.EntityFramework</Name>
+    <ProjectReference Include="..\BookStore.Data\BookStore.Data.csproj">
+      <Project>{a12af44b-904d-413b-aab8-93138e2526d6}</Project>
+      <Name>BookStore.Data</Name>
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
