+++ BookStore/BookStore.Data.Database/packages.config
