+++ BookStore/BookStore.Data/App.config
@@ -10,4 +10,7 @@
       <provider invariantName="System.Data.SqlClient" type="System.Data.Entity.SqlServer.SqlProviderServices, EntityFramework.SqlServer" />
     </providers>
   </entityFramework>
+  <connectionStrings>
+    <add name="BookStoreDb" connectionString="Data Source=LINKPC\SQLEXPRESS;Initial Catalog=;Integrated Security=True" providerName="System.Data.SqlClient" />
+  </connectionStrings>
 </configuration>
