--- BookStore/BookStore.Data/IAuthor.cs
@@ -1,10 +1,0 @@
-﻿using ZKorsakas.Data;
-
-namespace BookStore.Data
-{
-    public interface IAuthor : IEntity
-    {
-        string Name { set; get; }
-        string Surname { set; get; }
-    }
-}
