--- BookStore/BookStore.Data/IBook.cs
@@ -1,10 +1,0 @@
-﻿using ZKorsakas.Data;
-
-namespace BookStore.Data
-{
-    public interface IBook : IEntity
-    {
-        string Name { set; get; }
-        int Pages { set; get; }
-    }
-}
