+++ BookStore/BookStore.WebUI/Web.config
@@ -65,6 +65,6 @@
     </providers>
   </entityFramework>
   <connectionStrings>
-    <add name="BookStoreDb" connectionString="Data Source=LINKPC\SQLEXPRESS;Initial Catalog=BookStore.Data.EntityFramework.Contexts.BookStoreContext;Integrated Security=True" providerName="System.Data.SqlClient" />
+    <add name="BookStoreDb" connectionString="Data Source=LINKPC\SQLEXPRESS;Initial Catalog=;Integrated Security=True" providerName="System.Data.SqlClient" />
   </connectionStrings>
 </configuration>
